+++ packages/react/src/atomic/Select/src/Select.tsx
@@ -20,6 +20,7 @@ const Select = forwardRef<HTMLSelectElement, SelectBaseProps>(
       children,
       appearance = "neutral",
       aiGenerated = false,
+      size = "medium",
       ...rest
     },
     ref
@@ -30,6 +31,7 @@ const Select = forwardRef<HTMLSelectElement, SelectBaseProps>(
       </div>
       <select
         {...rest}
+        size={typeof size === "number" ? size : undefined}
         ref={ref}
         id={id}
         name={name}
@@ -38,6 +40,7 @@ const Select = forwardRef<HTMLSelectElement, SelectBaseProps>(
           aiGenerated
             ? select.classnames.aiGenerated
             : select.classnames.appearance[appearance],
+          select.classnames.size[typeof size === "number" ? "medium" : size],
         ]
           .filter(Boolean)
           .join(" ")}
@@ -48,7 +51,7 @@ const Select = forwardRef<HTMLSelectElement, SelectBaseProps>(
   )
 ) as ForwardRefExoticComponent<
   SelectBaseProps &
-    React.InputHTMLAttributes<HTMLSelectElement> &
+    Omit<React.InputHTMLAttributes<HTMLSelectElement>, "size"> &
     React.RefAttributes<HTMLSelectElement>
 > &
   SelectComponents;